Moves main.py to metaparticle.py, for import naming
