+++ python/metaparticle.py
